backend/requirements.txt
# FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Google APIs
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Database
pymysql==1.1.0
sqlalchemy==2.0.23
alembic==1.13.1

# Document Processing
python-docx>=0.8.11
PyPDF2>=3.0.0
pdfplumber==0.10.3

# Core dependencies
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.0.0

# Utilities
packaging>=21.0
cachetools>=4.2.0
rsa>=4.7

# OCR Support
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0

# HTTP Client
httpx==0.25.1

# JSON
orjson==3.9.10